refactor(util): move randomNumber to util package
