+++ src/main/java/racingcar/util/RandomNumber.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 public interface RandomNumber {
     int randomGenerate(int startInclusive, int endInclusive);
